+++ packages/widgets-manager/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Hotfix [2.2.1] - 2020-01-05
+### Fixed
+- Disable the Sync of firebase with redux -> Make a TODO to figure out corner
+cases and enable in the future
+
 ## Hotfix [2.2.0] - 2019-12-12
 ### Fixed
 - Save the initial_data { appSlug: {} } for each app in the store